mito-ai: address review
