+++ mito-ai/mito_ai/settings/handlers.py
@@ -3,9 +3,7 @@ import json
 from typing import Any, Final
 import tornado
 import os
-from mito_ai.utils.schema import MITO_FOLDER
-
-SETTINGS_PATH: Final[str] = os.path.join(MITO_FOLDER, 'settings.json')
+from mito_ai.settings.utils import SETTINGS_PATH, get_settings_field, set_settings_field
 
 
 class SettingsHandler(tornado.web.RequestHandler):
@@ -17,10 +15,6 @@ class SettingsHandler(tornado.web.RequestHandler):
         if not os.path.exists(SETTINGS_PATH):
             with open(SETTINGS_PATH, 'w') as f:
                 json.dump({}, f, indent=4)
-        
-    def check_xsrf_cookie(self) -> None:
-        """Override to disable CSRF protection for this handler."""
-        pass
     
     def get(self, key):
         """Get a specific setting by key"""
@@ -43,20 +37,3 @@ class SettingsHandler(tornado.web.RequestHandler):
         self.finish(json.dumps({"status": "updated", "key": key, "value": data['value']}))
 
 
-def set_settings_field(field: str, value: Any) -> None:
-    """
-    Updates the value of a specific field in settings.json
-    """
-    with open(SETTINGS_PATH, 'r') as user_file_old:
-        old_user_json = json.load(user_file_old)
-        old_user_json[field] = value
-        with open(SETTINGS_PATH, 'w+') as f:
-            f.write(json.dumps(old_user_json))
-            
-            
-def get_settings_field(field: str) -> Any:
-    """
-    Retrieves the value of a specific field from settings.json
-    """
-    with open(SETTINGS_PATH, 'r') as user_file_old:
-        return json.load(user_file_old).get(field)
