+++ mito-ai/src/RestAPI.tsx
@@ -2,6 +2,15 @@ import { PageConfig } from "@jupyterlab/coreutils";
 
 const baseUrl = PageConfig.getBaseUrl();
 
+export const getXsrfToken = (): string | null | undefined => {
+    const cookies = document.cookie.split(';');
+    const xsrfCookie = cookies.find(cookie => cookie.trim().startsWith('_xsrf='));
+    if (xsrfCookie) {
+        return xsrfCookie.split('=')[1];
+    }
+    return null;
+}
+
 /************************************
 
 SETTINGS ENDPOINTS
@@ -12,6 +21,7 @@ export const getSetting = async(settingsKey: string): Promise<string | undefined
     const response = await fetch(`${baseUrl}mito-ai/settings/${settingsKey}`, {
         headers: {
             'Content-Type': 'application/json',
+            'X-XSRFToken': getXsrfToken() || '',
         }
     });
     const data = await response.json();
@@ -23,6 +33,7 @@ export const updateSettings = async(settingsKey: string, settingsValue: string):
         method: 'PUT',
         headers: {
             'Content-Type': 'application/json',
+            'X-XSRFToken': getXsrfToken() || '',
         },
         body: JSON.stringify({ value: settingsValue }),
     });
